+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- New `bake` command line option (`-v` or `--product-version`) to build a specific product version ([#1])
+- New `bake` command line option `-v` (or `--product-version`) to build a specific product version ([#1])
+- New `bake` command line options `-q` (or `--product-version-quantile`) and `-m` (or `--max-bake-runners`) ([#1])
 
 [#1]: https://github.com/stackabletech/image-tools/pull/1